SS-54: refactored proto/grpc files naming
