+++ gradle.properties
@@ -1,0 +1,1 @@
+com.jules.ss.proto-files.version=1.0.15
