+++ src/main/java/com/ynero/ss/device/persistence/DeviceServiceMongoImpl.java
@@ -1,7 +1,7 @@
-package com.ynero.ss.device_service.persistence;
+package com.ynero.ss.device.persistence;
 
-import com.ynero.ss.device_service.domain.Device;
-import com.ynero.ss.device_service.domain.Port;
+import com.ynero.ss.device.domain.Device;
+import com.ynero.ss.device.domain.Port;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Primary;
