+++ src/main/java/com/ynero/ss/device/services/sender/PipelinesgRPCSender.java
@@ -1,21 +1,20 @@
-package com.ynero.ss.device.sender;
+package com.ynero.ss.device.services.sender;
 
-
-import com.ynero.ss.device_service.proto.main.grpc.PipelineRequestGrpc;
-import com.ynero.ss.device_service.proto.main.java.PipelinesMessage;
+import com.ynero.ss.pipeline.dto.proto.PipelinesMessage;
+import com.ynero.ss.pipeline.grpc.PipelineQueryReceiverServiceGrpc;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import org.springframework.stereotype.Service;
 
 @Service
-public class PipelinesgRPCSender{
+public class PipelinesgRPCSender {
 
     public void send(PipelinesMessage.PipelineQuery request) {
         ManagedChannel channel = ManagedChannelBuilder.forAddress("localhost", 9098)
                 .usePlaintext()
                 .build();
-        PipelineRequestGrpc.PipelineRequestBlockingStub stub = PipelineRequestGrpc.newBlockingStub(channel);
-        stub.receive(request);
+        var receiverServiceGrpcBlockingStub = PipelineQueryReceiverServiceGrpc.newBlockingStub(channel);
+        receiverServiceGrpcBlockingStub.receive(request);
         channel.shutdown();
     }
 }
