+++ src/test/java/com/ynero/ss/device/DeviceServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.ynero.ss.device_service;
+package com.ynero.ss.device;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
